P1 and P2 Controls solidified
Next is
>to set up a "Submit" function for selecting cards
>to set up an array of chosen cards

Let the Array of Chosen Cards be set in the
CHOSEN class.

The CHOSEN class attaches to a gameobject. It is an array of 10 Cards and 10 Buttons, associated by their position of the array. 0 is player 1. 1 ia player 2.

There will be two CHOSEN gameobjects in the scene, going into the serialize field of  DisplayHUD.

They will be in the bottom corners of the scene.

...actually, can we make these 10 buttons... 10 gameobjects with the tag: CHOSEN? and apply our GloveMovement's verification method to these 10 gameobjects?